[refactor/#5] data:user, data:user-api 모듈 분리
