+++ core/data/src/main/java/com/qriz/app/core/data/di/RepositoryModule.kt
@@ -2,8 +2,6 @@ package com.qriz.app.core.data.di
 
 import com.qriz.app.core.data.repository.OnBoardRepository
 import com.qriz.app.core.data.repository.OnBoardRepositoryImpl
-import com.qriz.app.core.data.repository.UserRepository
-import com.qriz.app.core.data.repository.UserRepositoryImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -15,13 +13,7 @@ import javax.inject.Singleton
 internal interface RepositoryModule {
     @Binds
     @Singleton
-    fun bindsUserRepository(
-        userRepositoryImpl: UserRepositoryImpl
-    ): UserRepository
-
-    @Binds
-    @Singleton
     fun bindsOnBoardRepository(
-       onBoardingRepository: OnBoardRepositoryImpl
+        onBoardingRepository: OnBoardRepositoryImpl
     ): OnBoardRepository
 }
