+++ core/data/user/user/.gitignore
@@ -1,0 +1,1 @@
+/build
