+++ feature/splash/src/main/java/com/qriz/app/feature/splash/SplashViewModel.kt
@@ -2,8 +2,8 @@ package com.qriz.app.feature.splash
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.qriz.app.core.data.repository.UserRepository
 import com.qriz.app.feature.splash.model.SplashEffect
+import com.quiz.app.core.data.user.user_api.repository.UserRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.FlowPreview
 import kotlinx.coroutines.flow.SharedFlow
