+++ settings.gradle.kts
@@ -33,3 +33,5 @@ include(":feature:splash")
 include(":feature:sign")
 include(":core:navigation")
 include(":feature:onboard")
+include(":core:data:user:user-api")
+include(":core:data:user:user")
